requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.39.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.2.0
pydantic>=2.6.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
jupyter>=1.0.0
ipykernel>=6.29.0
